+++ api/lib/application/organizations/index.js
@@ -454,30 +454,6 @@ const register = async function (server) {
       },
     },
     {
-      method: 'DELETE',
-      path: '/api/organizations/{id}/invitations/{organizationInvitationId}',
-      config: {
-        pre: [
-          {
-            method: securityPreHandlers.checkUserIsAdminInOrganization,
-            assign: 'isAdminInOrganization',
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            id: identifiersType.organizationId,
-            organizationInvitationId: identifiersType.organizationInvitationId,
-          }),
-        },
-        handler: organizationController.cancelOrganizationInvitation,
-        tags: ['api', 'invitations', 'cancel'],
-        notes: [
-          "- **Cette route est restreinte aux utilisateurs authentifiés en tant qu'admin d'une organisation**\n" +
-            "- Elle permet à l'administrateur de l'organisation d'annuler une invitation envoyée mais non acceptée encore.",
-        ],
-      },
-    },
-    {
       method: 'GET',
       path: '/api/organizations/{id}/member-identities',
       config: {
