+++ api/tests/acceptance/application/organizations/cancel-organization-invitation-delete_test.js
@@ -9,35 +9,6 @@ import {
 } from '../../../test-helper.js';
 
 describe('Acceptance | Route | Organizations', function () {
-  describe('DELETE /api/organizations/{id}/invitations/{invitationId}', function () {
-    it('should return 204 HTTP status code', async function () {
-      // given
-      const server = await createServer();
-
-      const { adminUser, organization } = await insertOrganizationUserWithRoleAdmin();
-      const invitation = databaseBuilder.factory.buildOrganizationInvitation({
-        organizationId: organization.id,
-        status: OrganizationInvitation.StatusType.PENDING,
-      });
-
-      const options = {
-        method: 'DELETE',
-        url: `/api/organizations/${organization.id}/invitations/${invitation.id}`,
-        headers: {
-          authorization: generateValidRequestAuthorizationHeader(adminUser.id),
-        },
-      };
-
-      await databaseBuilder.commit();
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(204);
-    });
-  });
-
   describe('DELETE /api/admin/organizations/{organizationId}/invitations/{invitationId}', function () {
     it('should return 204 HTTP status code', async function () {
       // given
