+++ api/tests/unit/application/organizations/index_test.js
@@ -509,27 +509,4 @@ describe('Unit | Router | organization-router', function () {
       expect(response.result.data).to.deep.equal([]);
     });
   });
-
-  describe('DELETE /api/organizations/{id}/invitations/{invitationId}', function () {
-    it('should call the cancel organization invitation controller', async function () {
-      // given
-      sinon
-        .stub(organizationController, 'cancelOrganizationInvitation')
-        .callsFake((request, h) => h.response('ok').code(200));
-      sinon.stub(securityPreHandlers, 'checkUserIsAdminInOrganization').returns(true);
-
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      const method = 'DELETE';
-      const url = '/api/organizations/1/invitations/1';
-
-      // when
-      await httpTestServer.request(method, url);
-
-      // then
-      expect(securityPreHandlers.checkUserIsAdminInOrganization).to.have.be.called;
-      expect(organizationController.cancelOrganizationInvitation).to.have.been.calledOnce;
-    });
-  });
 });
